feat: Add comment template to Trello cards
This commit introduces a new feature where a comment template is added to each Trello card. The template is loaded from a markdown file and added as a comment to the card when it is created. The path to the markdown file is specified in the settings.ini file. The run_trello_script.yml workflow