fix: update COMMENT_MD_PATH in settings.ini
This commit changes the value of COMMENT_MD_PATH in the settings.ini file. Previously, it was set to './path_to_your_md_file.md', but it has now been updated to 'comment.md'. This change will ensure that the application looks for the comment.md file in the correct location.